Cont. Responsive Dev on About page
Continued with responsive dev of about page.

Reached a minor issue and considering creating a branch to re-structure the layout of some of my HTML on the about page to see if this clears up the issue and makes the to be continued responsive dev a much more seamless proccess.