Added effect for a less harsher transition between page loads
Added a div to match the background of the current page which takes up the whole screen and then disappears upon the body fully loading.

It makes the transition between pages a little easier on the eye as I was having half a second delays between pages and it was a harsh flip from one bg color to another, namely when I moved from one of the core pages to a specific case study.

I also added some new content (images and videos) for some case studies.